ethniceats_backend/requirements.txt
django == 5.0
djangorestframework == 3.15.0
django-cors-headers
djangorestframework-simplejwt
tweepy
python-dotenv